+++ messages/ru-RU/app/modules/comments.php
@@ -4,7 +4,6 @@ return [
     'Comments' => "Комментарии",
     'Tree comments system' => "Древовидная система комментариев",
 
-
     'No comments available' => "Нет доступных для отображения комментариев",
     'An error occurred while retrieving the comments list' => "Произошла ошибка при получении списка комментариев",
 
@@ -13,6 +12,20 @@ return [
     '&larr; Prev page' => '&larr; Предыдущая страница',
     'Next page &rarr;' => 'Следующая страница &rarr;',
 
+    'ID' => "Просмотры",
+    'Parent ID' => "ИД родителя",
+    'Context' => "Контекст",
+    'Target' => "Цель",
+    'Name' => "Имя",
+    'E-mail' => "Эл. почта",
+    'Comment' => "Комментарий",
+    'User' => "Пользователь",
+    'User ID' => "ИД пользователя",
+    'Status' => "Статус",
+    'Session' => "Сессия",
+    'Created' => "Создано",
+    'Updated' => "Обновлено",
+    
 ];
 
 ?>